requirements.txt
channels == 4.2.2
Django == 4.2
django-environ == 0.12.0
djangorestframework == 3.16.0